requirements.txt
feedparser==6.0.11
jinja2==3.1.4
requests==2.32.3
pyyaml==6.0.2
